creative_studio_mcp/pyproject.toml
[project]
name = "creative-studio-mcp"
version = "0.1.0"
description = "MCP server for creative video production using FFmpeg"
dependencies = [
    "mcp>=1.0.0",
    "ffmpeg-python>=0.2.0"
]
requires-python = ">=3.10"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
